% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArboMAP_package_form.R
\name{ArboMAP-package}
\alias{ArboMAP-package}
\title{ArboMAP: Arbovirus Modeling and Prediction to Forecast Mosquito-Borne Disease Outbreaks}
\description{
Predict mosquito-borne disease outbreaks using distributed lags approach.
}
\details{
Code developed by JKD and MCW, Geography and Environmental Sustainability, University of Oklahoma.
ACK reformatted the code to be in package format.
}
\author{
Justin K. Davis, Michael C. Wimberly (justinkdavis@ou.edu, mcwimberly@ou.edu), Alexander C. Keyel (akeyel@albany.edu)
}
