% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArboMAP_package_form.R
\name{read.weather.data}
\alias{read.weather.data}
\title{Read Weather Data (ArboMAP)}
\usage{
read.weather.data(weatherpathstr, weathersummaryfile)
}
\arguments{
\item{weatherpathstr}{A directory containing the weather data to be aggregated.
The weather data can contain multiple files, and these can overlap in date range,
as long as they all have the same columns.
The weather data files in this folder should have a district column, a doy (day of year) column
a year column, and then an entry for each independent variable of interest}

\item{weathersummaryfile}{If this file is in weatherpathstr, it will be ignored.
Note that no consolidated summary file is created.}
}
\description{
Read in the ArboMAP formatted weather data. Data in county-year format for the RF1 model
should be added using the files.to.add entry in the rf1.inputs object.
}
\details{
Data in this format can be downloaded from Google Earth Engine, see ArboMAP documentation
for instructions

Written by J. Davis and M. Wimberly. Modified by A. Keyel. Original available at www.github.com/ecograph/ArboMAP. Used under GPL-3 license
}
